Fix JS error that appears when scrolling on pages
I created an event listener that changes the "active" class for
the table of contents items depending on how far you are down on
a page. However, it was also triggering on Pages, which don't have
a table of contents, and throwing an error.

We only want this to apply for articles, not for pages.

Added an if block as a sanity check, and removed the script line
itself from the "page" templates.